+++ tsup.config.ts
@@ -1,0 +1,34 @@
+import { defineConfig, type Options } from "tsup";
+
+const baseOptions: Options = {
+  clean: true,
+  dts: true,
+  entry: ["src/index.ts"],
+  minify: true,
+  skipNodeModulesBundle: true,
+  sourcemap: true,
+  target: "es6",
+  tsconfig: "./tsconfig.json",
+  keepNames: true,
+  cjsInterop: true,
+  splitting: true,
+};
+
+export default [
+  defineConfig({
+    ...baseOptions,
+    outDir: "dist/cjs",
+    format: "cjs",
+  }),
+  defineConfig({
+    ...baseOptions,
+    outDir: "dist/esm",
+    format: "esm",
+  }),
+  defineConfig({
+    ...baseOptions,
+    outdir: "dist",
+    format: "iife",
+    globalName: "protomapsL"
+  })
+];